clean code and refactor pages
